requirements.txt
rich
pyfiglet

socket
scapy
nmap
ipaddress
dns.resolver
openai
shodan
vulners


threading
time
random
requests
concurrent
pyttsx3
pyler
datetime

pathlib
json